requirements.txt
quart>=0.18.4
quart-cors>=0.6.0
geopy>=2.3.0
scikit-learn>=1.3.0
httpx>=0.24.0
python-dotenv>=1.0.0
openai>=1.0.0
